+++ cerebellar_models/connectome/io_purkinje.py
@@ -22,12 +22,12 @@ class DuplicateSynapses(AfterConnectivityHook):
     def postprocess(self):
         for cs_name in self.conn_strategy.get_output_names():
             # Draw the number of connection to create
-            n = int(self.contacts.draw(1)[0])
             cs = self.scaffold.get_connectivity_set(cs_name)
             pre_ps = cs.pre_type.get_placement_set()
             post_ps = cs.post_type.get_placement_set()
-            if n > 1:
-                pre_locs, post_locs = cs.load_connections().all()
-                pre_locs = np.repeat(pre_locs, n - 1, axis=0)
-                post_locs = np.repeat(post_locs, n - 1, axis=0)
-                cs.connect(pre_ps, post_ps, pre_locs, post_locs)
+            duplicates = np.asarray(self.contacts.draw(len(cs)), dtype=int)
+            for (pre_locs, post_locs), n in zip(cs.load_connections(), duplicates):
+                if n > 1:
+                    pre_locs = np.repeat([pre_locs], n - 1, axis=0)
+                    post_locs = np.repeat([post_locs], n - 1, axis=0)
+                    cs.connect(pre_ps, post_ps, pre_locs, post_locs)